requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
pandas==2.2.0
mysql-connector-python==8.3.0
sqlalchemy==2.0.27
matplotlib==3.8.2
seaborn==0.13.2
wordcloud==1.9.3
nltk==3.8.1
streamlit==1.31.1
plotly==5.18.0
